# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestão de imóveis para locação",
        "title": "InquilinoTop API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/leases": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Lista contratos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Cria contrato",
                "parameters": [
                    {
                        "description": "Dados do contrato",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lease.CreateLeaseInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/leases/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Busca contrato por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do contrato",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Atualiza contrato",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do contrato",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do contrato",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lease.UpdateLeaseInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leases"
                ],
                "summary": "Remove contrato (soft-delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do contrato",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/leases/{leaseId}/payments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Lista pagamentos de um contrato",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do contrato",
                        "name": "leaseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Registra pagamento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do contrato",
                        "name": "leaseId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do pagamento",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment.CreatePaymentInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/payments/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Atualiza pagamento (ex: marcar como pago)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do pagamento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do pagamento",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment.UpdatePaymentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "lease.CreateLeaseInput": {
            "type": "object",
            "properties": {
                "deposit_amount": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "rent_amount": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "unit_id": {
                    "type": "string"
                }
            }
        },
        "lease.UpdateLeaseInput": {
            "type": "object",
            "properties": {
                "deposit_amount": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "rent_amount": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "payment.CreatePaymentInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "due_date": {
                    "type": "string"
                },
                "lease_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "payment.UpdatePaymentInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "paid_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token no formato: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}